feat: added modelId to file upload table
