+++ packages/server/modules/core/dbSchema.ts
@@ -498,6 +498,7 @@ export const FileUploads = buildTableHelper('file_uploads', [
   'streamId',
   'branchName',
   'userId',
+  'modelId',
   'fileName',
   'fileType',
   'fileSize',
