+++ packages/server/modules/fileuploads/domain/operations.ts
@@ -3,16 +3,22 @@ import {
   FileUploadRecord,
   FileUploadRecordV2
 } from '@/modules/fileuploads/helpers/types'
-import {
-  SaveUploadFileInput,
-  SaveUploadFileInputV2
-} from '@/modules/fileuploads/repositories/fileUploads'
 import { Optional } from '@speckle/shared'
 import { FileImportResultPayload } from '@speckle/shared/dist/commonjs/workers/fileimport/job.js'
 
 export type GetFileInfo = (args: {
   fileId: string
 }) => Promise<Optional<FileUploadRecord>>
+
+export type SaveUploadFileInput = Pick<
+  FileUploadRecord,
+  'streamId' | 'branchName' | 'userId' | 'fileName' | 'fileType' | 'fileSize'
+> & { fileId: string }
+
+export type SaveUploadFileInputV2 = Pick<
+  FileUploadRecordV2,
+  'projectId' | 'modelId' | 'userId' | 'fileName' | 'fileType' | 'fileSize'
+> & { fileId: string }
 
 export type SaveUploadFile = (args: SaveUploadFileInput) => Promise<FileUploadRecord>
