+++ packages/server/modules/fileuploads/helpers/types.ts
@@ -11,6 +11,7 @@ export type FileUploadRecord = {
   id: string
   streamId: string
   branchName: string
+  modelId: Nullable<string>
   userId: string
   fileName: string
   fileType: string
@@ -26,7 +27,7 @@ export type FileUploadRecord = {
 export type FileUploadRecordV2 = {
   id: string
   projectId: string
-  modelId: string
+  modelId: Nullable<string>
   userId: string
   fileName: string
   fileType: string
